voice_agent/requirements.txt
livekit-agents[openai,silero]==1.2.14
livekit-plugins-noise-cancellation==0.2.5
python-dotenv>=1.0.1
mcp[cli]>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.22.0
